+++ Sidequest3/Views/VisualEffectBlur.swift
@@ -1,0 +1,21 @@
+//
+//  VisualEffectBlur.swift
+//  Sidequest3
+//
+//  Created by ole on 30.03.26.
+//
+
+import SwiftUI
+
+struct VisualEffectBlur: UIViewRepresentable {
+    var blurStyle: UIBlurEffect.Style
+    
+    func makeUIView(context: Context) -> UIVisualEffectView {
+        let view = UIVisualEffectView(effect: UIBlurEffect(style: blurStyle))
+        return view
+    }
+    
+    func updateUIView(_ uiView: UIVisualEffectView, context: Context) {}
+}
+
+
